support the farmer plugin. implements #1
This makes adjustments so that farming will automatically work as soon
as the farmer plugin is installed. The setup step can be skipped and
instead the plugin is preconfigured to use /storage/farm as farm dir.

!bang based animals are supported, but of course it makes more sense to
use a host based setup with docker. With a reverse proxy in front of
this image, it should be easy to point different host names at this
image. However this has not been tested yet and might need additional
adjustments in the farmer plugin.

A merge should only happen after this has been tested.